Separated test files for distinction of system being tested
